refactor(react): reorganize exports
