+++ packages/react/src/ui/index.ts
@@ -1,2 +1,3 @@
 export * from './components';
 export * from './provider';
+export * from './theme';
