+++ packages/react/src/ui/provider/index.ts
@@ -1,2 +1,1 @@
-export * from './ThemePreferencesProvider';
 export * from './UiProvider';
