+++ packages/react/src/ui/theme/components/ThemeEditor.tsx
@@ -1,6 +1,6 @@
 import { Box, Button, Center, Flex, Modal, SegmentedControl, Title } from '@mantine/core';
 import { Icon } from '../../components';
-import { useThemePreferences } from '../../provider';
+import { useThemePreferences } from '../provider';
 import { ColorPicker } from './ColorPicker';
 
 interface ThemeEditorProps {
