+++ packages/react/src/ui/theme/components/index.ts
@@ -1,1 +1,2 @@
+export * from './ColorPicker';
 export * from './ThemeEditor';
