+++ packages/react/src/ui/theme/index.ts
@@ -1,2 +1,2 @@
 export * from './components';
-export * from './utils';
+export * from './provider';
